FInished backend

+++ DTO/PayToClientDTO.cs
@@ -5,9 +5,9 @@ public class PayDto
 {
         public int Id { get; set; }
         public required string Currency { get; set; }
-        public float MonthlyPayment { get; set; }
-        public float AnnualMultiplierPayment { get; set; }
-        public float InterestMonthlyPayment { get; set; }
+        public decimal MonthlyPayment { get; set; }
+        public decimal AnnualMultiplierPayment { get; set; }
+        public decimal InterestMonthlyPayment { get; set; }
 
         public int? SubscriptionId { get; set; }
 }
